initial verison 0x13.
this change is opaque to clients of the library. as far as they're
concerned, there is still just one endpoint (`Argon2::new`, which hashes
to version 0x13) and its derivatives.

however, the verifier needs to support both 0x10 and 0x13 encoded
strings, so we use pub restrict to restrict access to 0x10 Argon2
creation.